Refactor the Menu class
I realized all the menu option classes are basically just a string and an event, so this converts that into a struct that the Menu class has a list of.
