+++ SFMLGame/GameLoader.cpp
@@ -7,8 +7,6 @@
 #include "EventQueue.h"
 #include "GameClock.h"
 #include "InputReader.h"
-#include "BackMenuOption.h"
-#include "QuitMenuOption.h"
 #include "Menu.h"
 #include "PlayingStateInputHandler.h"
 #include "MenuStateInputHandler.h"
@@ -38,11 +36,7 @@ shared_ptr<Game> GameLoader::Load() const
    auto eventQueue = make_shared<EventQueue>();
    auto clock = shared_ptr<GameClock>( new GameClock( renderConfig ) );
    auto inputReader = shared_ptr<InputReader>( new InputReader( gameConfig ) );
-   auto backMenuOption = shared_ptr<BackMenuOption>( new BackMenuOption( eventQueue ) );
-   auto quitMenuOption = shared_ptr<QuitMenuOption>( new QuitMenuOption( eventQueue ) );
-   auto menu = make_shared<Menu>();
-   menu->AddOption( backMenuOption );
-   menu->AddOption( quitMenuOption );
+   auto menu = shared_ptr<Menu>( new Menu( eventQueue ) );
    auto playingStateInputHandler = shared_ptr<PlayingStateInputHandler>( new PlayingStateInputHandler( inputReader, gameConfig, eventQueue ) );
    auto menuStateInputHandler = shared_ptr<MenuStateInputHandler>( new MenuStateInputHandler( inputReader, eventQueue, menu ) );
    auto closingStateInputHandler = make_shared<ClosingStateInputHandler>();
