--- SFMLGame/IMenuOption.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#include "Common.h"
-
-NAMESPACE_BEGIN
-
-class __declspec( novtable ) IMenuOption
-{
-public:
-   virtual std::string GetText() const = 0;
-   virtual void Select() = 0;
-};
-
-NAMESPACE_END
