--- SFMLGame/QuitMenuOption.cpp
@@ -1,16 +1,0 @@
-#include "QuitMenuOption.h"
-#include "EventQueue.h"
-
-using namespace NAMESPACE;
-using namespace std;
-
-QuitMenuOption::QuitMenuOption( shared_ptr<EventQueue> eventQueue ) :
-   _eventQueue( eventQueue ),
-   _text( IDS_MenuOptionQuit )
-{
-}
-
-void QuitMenuOption::Select()
-{
-   _eventQueue->Push( GameEventType::Quit );
-}
